requirements.txt
absl-py==0.11.0
aniso8601==8.0.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-RESTful==0.3.8
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase64==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.3
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-ImageAttach==1.1.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.10.1
urllib3==1.25.11
Wand==0.6.3
Werkzeug==1.0.1
wrapt==1.12.1
